Separated DriveMotor from Motor
